+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,10 @@
         android:theme="@style/Theme.TTMatchLog"
         tools:targetApi="31">
         <activity
-            android:name=".view.LoginActivity"
+            android:name=".presentation.view.activity.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".view.SigninActivity"
+            android:name=".presentation.view.activity.SigninActivity"
             android:theme="@style/Theme.App.SplashScreen"
             android:exported="true">
             <intent-filter>
@@ -26,7 +26,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".view.MainActivity"
+            android:name=".presentation.view.activity.MainActivity"
             android:exported="true">
         </activity>
     </application>
