+++ app/src/main/res/layout/activity_login.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="30dp"
-    tools:context=".view.LoginActivity">
+    tools:context=".presentation.view.activity.LoginActivity">
 
     <EditText
         android:id="@+id/email"
