+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".view.MainActivity">
+    tools:context=".presentation.view.activity.MainActivity">
 
     <TextView
         android:id="@+id/userNameText"
